RankingMODEL/autoscout-ml/requirements_flask.txt
flask==3.0.3
flask-cors==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
numpy==1.24.3
gunicorn==21.2.0
